Protect all endpoints with jwt still think node is Confusing

+++ Controllers/contact-controller.js
@@ -9,9 +9,10 @@ const Contact = require("../models/contact")
  * @param {Express.Request} req 
  * @param {Express.Response} res 
  * @route GET /api/contact
+ * @access protected with JWT token
  */
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find()
+    const contacts = await Contact.find({ user_id: req.user.id })
     res.json({
         msg: "Contacts", method: req.method, contacts
     })
@@ -39,6 +40,7 @@ const getContact = asyncHandler(async (req, res) => {
  * @param {Express.Request} req 
  * @param {Express.Response} res 
  * @route POST /api/contact
+ * @access protected with JWT token
  */
 const createContact = asyncHandler(async (req, res) => {
     res.status(200)
@@ -48,7 +50,7 @@ const createContact = asyncHandler(async (req, res) => {
         throw new Error("All fields are mandatory !");
     }
     const newContact = await Contact.create({
-        name, email, phone
+        name, email, phone, user_id: req.user.id
     })
     res.json({
         msg: "Contacts", method: req.method, contact: newContact
@@ -60,8 +62,15 @@ const createContact = asyncHandler(async (req, res) => {
  * @param {Express.Request} req 
  * @param {Express.Response} res 
  * @route PUT /api/contact/:id
+ * @access protected with JWT token
  */
 const updateContact = asyncHandler(async (req, res) => {
+    const contact = await Contact.findById(req.params.id)
+    if (contact.id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not Authorized")
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id,
         req.body, {
         new: true
@@ -82,9 +91,14 @@ const updateContact = asyncHandler(async (req, res) => {
  * @param {Express.Request} req 
  * @param {Express.Response} res 
  * @route DELETE /api/contact/:id
+ * @access protected with JWT token
  */
 const deleteContact = asyncHandler(async (req, res) => {
-
+    const contact = await Contact.findById(req.params.id)
+    if (contact.id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Not Authorized")
+    }
     const deletedContact = await Contact.findByIdAndRemove(req.params.id)
     if (!deletedContact) {
         res.status(404)
